orchestrator/requirements.txt
fastapi
uvicorn
httpx
PyGithub
google-genai
mcp
pytest
pytest-asyncio
langchain
langgraph
langchain-google-genai
opentelemetry-api
opentelemetry-sdk
